rUM: R Templates from the University of Miami

This holds some r markdown and quarto templates and a template to create a research project in "R Studio".

See also

Author

Maintainer: Raymond Balise balise@miami.edu (ORCID)

Authors:

Gabriel Odom gabriel.odom@fiu.edu (ORCID)

Kyle Grealis kxg679@miami.edu (ORCID)

Francisco Cardozo foc9@miami.edu (ORCID)

Other contributors:

Frank Gutierrez frank14898@outlook.com (ORCID) [contributor]